refactor step
