+++ models/Space.js
@@ -24,7 +24,7 @@ exports.create = function(req, res) {
 
 exports.fetch = function(req, res) {
   Space.run().then(function(spaces) {
-    res.render('spaces/index', { title: 'Spaces', spaces: spaces });
+    res.render('spaces/index', { title: 'Spaces', spaces: spaces,  user: req.session.object });
   })
 
 };
