+++ routes/spaces.js
@@ -4,7 +4,8 @@ var space = require('../models/Space');
 
 
 router.get('/', function(req, res, next) {
-  res.render('spaces/index', { title: 'Spaces', user: req.session.object });
+  space.fetch(req, res);
+  // res.render('spaces/index', { title: 'Spaces', user: req.session.object });
 });
 
 router.get('/new', function(req, res, next) {
